enemy default + status effect buff
enemy default + status effect buff
