+++ Assets/Enemy/EnemyDefault.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: f36ad407132336c459de1fe86b5e3f0b
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
