+++ Assets/Enemy/Snail Juggernaut.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 6dfd2c2a4a3561b42b5365b310bca322
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
