+++ Assets/PeriodicAStarMovement.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b1cc39a131121604d9543fcb61b44e3e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
