+++ Assets/PowerUps/StatusEffectDurationPU.asset
@@ -1,0 +1,19 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 3ad1b7eee338ced468afe1acdd8e145c, type: 3}
+  m_Name: StatusEffectDurationPU
+  m_EditorClassIdentifier: 
+  powerUpName: Longer Effects
+  powerUpDescription: Stun/Slow enemies for longer
+  powerUpColor: {r: 0.3112822, g: 0.6415094, b: 0, a: 1}
+  sprite: {fileID: 21300000, guid: 23eacd77299d67f498e85858076a33f2, type: 3}
+  durationIncreaseAmount: 0.1
