+++ Assets/PowerUps/StatusEffectDurationPU.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8ccc30f21d67bf34f94269f69ba16db3
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
