+++ Assets/PowerUps/StatusEffectDurationUp.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3ad1b7eee338ced468afe1acdd8e145c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
