+++ Assets/PowerUps/UI_Option.prefab
@@ -35,7 +35,7 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0.5, y: 0.5}
   m_AnchorMax: {x: 0.5, y: 0.5}
-  m_AnchoredPosition: {x: 0, y: 28.4}
+  m_AnchoredPosition: {x: 0, y: -4.4}
   m_SizeDelta: {x: 129.23, y: 32.5424}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!222 &2463478562823524909
