+++ Assets/StatusEffect/StatusMarker.prefab
@@ -73,7 +73,7 @@ SpriteRenderer:
   m_SortingLayerID: 0
   m_SortingLayer: 0
   m_SortingOrder: 3
-  m_Sprite: {fileID: 21300000, guid: cd623124aae2860459172810a616b806, type: 3}
+  m_Sprite: {fileID: 21300000, guid: e2d60e25d1677c64aad0d7b9cb662cf5, type: 3}
   m_Color: {r: 1, g: 1, b: 1, a: 1}
   m_FlipX: 0
   m_FlipY: 0
